Game Design Implemented
Based off application of the pigeonhole theorem to digital sums, where a random number of three digit numbers are generated, and that number is divided by the highest possible number of unique digital sums derived from a three digit number. The floor of this operation is one output (the maximum repeats for every possible digital sum) and the remainder is the other output (the maximum number of digital sums that are repeated more than the others). Conditions for successful outcomes are still to be determined (maybe based on even and odd combinations of common- and lone- repeats?).